# api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Kurator API",
    "version": "1.0",
    "description": "REST API for Kurator: collections, items, search, external metadata lookup, session-based auth, bundled SQL migrations with first-boot bootstrap, and optional S3-backed image uploads.\n\n**API v2:** **`/api/v2/hitlists`** adds permalink slugs, internet-**public** visibility, stub entries, votes, and comments (see tag hitlists-v2).\n\n**Authentication:** Sessions are opaque tokens stored hashed in the database. The API accepts (1) the **kurator_session** HTTP-only cookie (typical browser flow via the Next.js same-origin proxy) or (2) **`Authorization: Bearer <session_token>`** on any protected route. When both are present, the **cookie wins**. After **POST /api/v1/auth/register**, **POST /api/v1/auth/login** (when 2FA is not required), or **POST /api/v1/auth/login/2fa**, the JSON body includes **session_token** \u2014 the same secret as the cookie \u2014 for native and non-cookie clients. **POST /api/v1/auth/logout** revokes the session for either mechanism; browsers also receive `Set-Cookie` clearing the cookie."
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
...
  "securityDefinitions": {
    "SessionCookie": {
      "type": "apiKey",
      "name": "kurator_session",
      "in": "cookie",
      "description": "HTTP-only session cookie set on successful register/login/2fa. Preferred when both cookie and Authorization Bearer are present."
    },
    "BearerToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer followed by the opaque session_token from POST /auth/login, /auth/login/2fa, or /auth/register (same secret as the kurator_session cookie). Example: Bearer <session_token>"
    }
  },
  "paths": {
...
    "/api/v1/billing/create-checkout-session": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Stripe Checkout session",
        "description": "Requires **SessionCookie** or **BearerToken**. Returns a Stripe Checkout URL for Kurator Pro (monthly or annual).",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerToken": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateCheckoutSessionBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BillingSessionURLResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Already subscribed"
          },
          "503": {
            "description": "Billing not configured"
          }
        }
      }
    },
    "/api/v1/billing/switch-interval": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Switch Pro billing interval",
        "description": "Requires **SessionCookie** or **BearerToken**. Updates an active Pro subscription between monthly and annual Stripe prices (proration handled by Stripe).",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerToken": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SwitchBillingIntervalBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "No subscription or invalid interval"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Already on that interval or cannot switch"
          },
          "503": {
            "description": "Billing not configured"
          }
        }
      }
    },
    "/api/v1/billing/portal": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Stripe Customer Portal session",
        "description": "Requires **SessionCookie** or **BearerToken**. Returns a Stripe Customer Portal URL to manage an existing subscription.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerToken": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BillingSessionURLResponse"
            }
          },
          "400": {
            "description": "No Stripe customer"
          },
          "401": {
            "description": "Unauthorized"
          },
          "503": {
            "description": "Billing not configured"
          }
        }
      }
    },
    "/webhooks/stripe": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Stripe webhook",
        "description": "Verifies Stripe-Signature using STRIPE_WEBHOOK_SECRET. Handles checkout.session.completed, customer.subscription.created/updated/deleted, invoice.paid, and invoice.payment_failed.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Invalid payload or signature"
          },
          "503": {
            "description": "Billing not configured"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateCheckoutSessionBody": {
      "type": "object",
      "required": [
        "interval"
      ],
      "properties": {
        "interval": {
          "type": "string",
          "enum": [
            "monthly",
            "annual"
          ]
        }
      }
    },
...
    "SwitchBillingIntervalBody": {
      "type": "object",
      "required": [
        "interval"
      ],
      "properties": {
        "interval": {
          "type": "string",
          "enum": [
            "monthly",
            "annual"
          ]
        }
      }
    },
    "BillingSessionURLResponse": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        }
      }
    },
...
  }
}